cloud_functions/showcase_data_ingestion/requirements.txt
google-api-core==2.11.0
google-auth==2.18.1
google-cloud-core==2.3.2
google-cloud-documentai==1.4.2
google-cloud-firestore==2.5.3